requirements.txt
beautifulsoup4==4.4.1
click==6.6
cssselect==0.9.2
lxml==3.6.0
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
wheel==0.24.0
